fix: torch_dtype not dtype in from_pretrained (actual root cause)
AutoModelForCausalLM.from_pretrained() takes torch_dtype=, not dtype=.
The wrong kwarg silently passed on some transformers versions but correctly
raises TypeError on others. Fixes calibration crash:
  TypeError: Qwen2ForCausalLM.__init__() got an unexpected keyword argument 'dtype'

Also reverts the transformers<4.53 pin added in the previous commit — that
was chasing the wrong fix. Keeps trl>=0.16,<0.17 explicit.

+++ src/agentdebugger/training/grpo.py
@@ -384,7 +384,7 @@ def train(config: TrainingConfig) -> None:
             adapter_dir = str(stage_dir)
             continue
 
-        model = AutoModelForCausalLM.from_pretrained(config.model, device_map="auto", dtype=dtype)
+        model = AutoModelForCausalLM.from_pretrained(config.model, device_map="auto", torch_dtype=dtype)
         model.config.use_cache = False
         model = _build_lora_model(model, profile, adapter_dir)
         if stage_index == 0:
